requirements.txt
astroid==2.4.2
ffmpeg-python==0.2.0
future==0.18.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
Pillow==8.0.0
pygame==2.0.0.dev14
pylint==2.6.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
